pyproject.toml
[project]
name = "is-network-up"
version = "0.1.0"
description = "Network probe"
authors = [
    {name = "Ted Romer",email = "tromer@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-monitoring (>=2.26.0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
